Joing URLs now works like joining filenames: with __div__.
URLAttr no longer has a base attribute, but the base will be directy
incorporated into the URL when parsing. When publishing the URL will
again be published relative to the output filename for *all* URLs, not
just for root URLs.

As the file URL is no available during convert calls, calculating image
sizes is again done in xist.ns.specials.autoimg.publish.