Move the check for empty elements with content into a new method checkValid(). But the test is still done in append() and insert(). To be consistent with the attribute checks (required and values) this should be done only when parsing or publishing.
Frag, Attr and Attrs now have a new method clear(), that remove the
content.

Attr.convert() and other methods that created new attributes had the
following bug: The new attribute was created by calling the constructor
without arguments. This created an attribute that had the default value
and then the content of the old attribute was converted over, so the new
attribute always started with the default value. This has been fixed
by introducing a private method _create() in Frag, Attr and Attrs
that is responsible for creating an empty copy of self.

Several of the methods in Frag (mapped, sorted, etc.) explicitely
instanciated a Frag. This has been changed to using the new
_create() method, so these method work correctly for attributes
now.

Added a __repr__ to Attr.__metaclass__ that mentions that this is
an *attribute* class.

Merged in the release 1.5.3 which add ll.xist.ns.php.expression.

Updated the test to work again with the two new specials modules
and the xist entity which has moved to abbr.

Updated the test to check the functionality of default attributes
(assigning, deleting etc.) in the context of convert.