Renamed the cols argument of asText and defaulted it to 72.
Now the class Context is nested inside node and can be overwritten.
Converter.__getitem__ will always instantiate the appropriate
Context class for the DOM class passed in, so now it's possible
to overwrite the Context constructor and add the attributes
required. ll.xist.ns.doc uses this new features and has the following
extended functionality: Now not only the depth is recorded, but section
numbers are kept, so it would be possible to generate an automatic
section numbering.

Enhanced to HOWTO.