Removed Windows linefeeds from ll.xist.ns.ihtml.
ll.xist.ns.example now uses a nested title element just like
section does. Updated the HOWTO accordingly.

Added a note to the HOWTO that the convert method is const.

Added a note to the HOWTO that URLAttr.forInput() can be
used to get an URL suitable for reading/writing.

Replace hasAttr with hasattr everywhere.