Attr now supports the method filtered(). This is used by without() now. The arguments for without() have changed, because handling global attributes was too magical.
A new method with() has been added, with does
the opposite of without(), i.e. it removes all
attributes that are not specified as parameters.

The Python name of each Node subclass is now available
as the class attribute pyname. This should prepare us
for the change of nested class names in Python 2.3.

withSep() has been renamed to withsep().

Added several tests for with(), getns() and withsep().