Started rewrite of namespace handling: Now namespaces are classes and all element, procinst and entity classes are simply nested classes inside this namespace class. Derived namespaces will automatically define inherited element, procinst and entity classes. This makes it possible to overwrite element classes and have the other classes pick up those changed classes automatically. For an example see the method test_autoinherit in the test suite.
Conversion targets are now always namespace classes. The will now be
used in two ways: 1) The are used tith an issubclass test to determine
with conversion is wanted. 2) Once this is done, the target object is used
for instantiating the new converted tree, so it's possible to e.g. pass in
a namespace derived from html and overwrite the a element, so that it
always generates a target="_top" attribute and pass in this namespace as
the conversion target.

Names of element classes etc. are now kept in a tuple with both the Python
and XML name. This tuple replaces the old xmlname attribute. This makes
it possible to handle both Python and XML names in the same way. Many
of the attribute handling methods have now got an additional argument
xml, that specifies whether the XML or the Python name should be used.

The Element methods that only forward the call to the appropriate method
of the attrs attribute are now deprecated, as calling the method an
attrs is faster, and only two character more to type in most cases.

Added Andreas tld.py module and tld2xsc.py script, that can be
used for generating XIST namespace skeletons from Java Tag Library
Descriptor files.