Added missing multi_case element and auto_restore attribute.
Removed the inheritable attributes from wrapper again, as this will
be done differently.

Added DTD like information at the end of the namespace. This adds a content
attribute to all element classes, that specifies which other elements are
allowed in the content. This ist not complete yet, and will be removed again,
once the inheritable attributes are done.