Defining element nested inside a namespace class brings too many problems. All of these problems are connected with the fact that a class namespace is not part of Pythons nested scopes, so it's impossible to have to element classes inside a namespace class, where one element class is derived from another. As this will be quite frequent this approach was abandoned. Now element classes are again defined at the module level. Creating namespaces is still done by deriving for xsc.Namespace, but with any element classes nested inside. These element classes are later added by calling xsc.makensmod(nsobj, vars()), which also replaces the module in sys.modules with the namespace class.
For this the work it's now possible to assign element classes to
namespaces after the namespaces have been created.
xsc.Namespace.__metaclass__.__setattr__ will do all the bookkeeping
(with the help of the _registerns classmethod)

There are many namespace that still have to be converted back to the old
method (module level element classes) and several conversion method
have still to be converted back too.