When checking whether a file should be mentioned in the repr of a namespace. __file__ attributes from bases classes must be ignored.
CharRef is no longer derived from Entity as this lead to problems in the
metaclasses.

ll.xist.ns.xsc now has attributes _elements, _procinsts, _entities
and _charrefs that work just like in attributes of the same name
in Namespace classes, but contain those elements etc. that are
outside of any namespace. The plan is to make those accessible
with methods like iterelementkeys() etc. just like for element
within a namespace. Whether a new special namespace will be
introduced for that is not clear yet.