The conv() has two new arguments: function and node
function (when set to a callable object) will be used for
calling the method mapped() instead of calling convert().
In addition to that this callable object will be stored
in the converter, which must be passed around all calls to mapped().

node will be set to self by conv() automatically. This makes it
possible to access the "document root" during conversion.

Update the test script.