Update NEWS.
Add methods updateexisting() and updatenew() to Namespace.

Removed the default namespace, as this was problematic. In the following
example, foo.xmlns.element("foo") would raise an error:

from ll.xist import xsc

def f():
   class xmlns(xsc.Namespace):
      class foo(xsc.Element):
         pass

class foo(xsc.Element):
   pass

f()

print foo.xmlns.element("foo")

The reason for this is that the global foo is assigned to the default
namespace. Then f is called and during the creation of the namespace
class the inner class foo is temporarily assigned to the global namespace
overwriting the global foo class. What's missing is a way to access all
global element classes, but this might be implemented via __subclasses__.