Update requirements in INSTALL.xml.
Use the _makestring() method in more spots in the parser, this might
allow us to use XMLUnicodeParser someday in the future.

Now that Python 2.3 is required use the warnings module directly.

Rewrote handling of entity and character references. Now both
SGMLOPParser and BadEntityParser should work correctly. Added tests
to verify that.