Add a new way (named XFind) to iterate through trees, that looks like XPath: You can do node/filter or iterator/filter or node/filter/filter etc.
Change walk() to support XFind: no you can do node.walk()/html.a.

Add a class NSPool that is a pool of namespaces. This will be used by
the parser when it encounters xmlns attributes to get a namespace
class for a namespace name.

Add notes in xsc.py where it would make sense to switch to generator
expression in Python 2.4.

Update examples to use XFind.