The xfind __getitem__ now supports negative indizes (by using a queue) and slices (by converting the content to a list).
Fixed the bug that a/b/c didn't work any more: Iterating a/b must resolve
the rest of the chain explicitely. Returning an _XFinder object for the
rest of the chain isn't enough.

Pass arguments to xfind() as xfind(iterator, *operators) instead of
xfind(iterator, operators).

Fix regression in _XFinder.__repr__.