Begin work on removing tidy stuff and switching to the HTML parser from libxml2.
Provide the option for the parser to ignore illegal elements, attribute,
processing instructions and entity references. All the methods in Prefixes
that return the appropriate classes will now issue a warning and return None
if an error occurs. The default warning handling for those warnings is to
raise an exception. The parser is able to continue after the error, if the
warning doesn't result in an exception. A second set of warning has been
defined which will be used during parsing. All these warning classes have a
common base class named ParseWarning. This makes it possible to silence all
these warnings with one call to warnings.filterwarnings.