Rename methos a bit: element_keys_pi() is elementkeys() and element_keys_xml() is elementkeys_xml(). element_py is elementclass() and create_element_py() is element(). Do a similar renaming for attribute methods (and drop the xml parameter here too).
Remove test for xmlprefix() method (which is gone).

Skip struts_html.taglib in test_namespaces, as this would require a prefix,
which isn't generated automatically, as taglib is a procinst not an element.

Collapse list and iter version of attribute iteration methods into one:
e.g. instead of iteralloweditems() and alloweditems() there's only
alloweditems() now (which behaves like iteralloweditems() (similar to the
changes for dict.keys() etc. planned for Py3).

Implement the foo/foo_xml split for xfind.hasattrnamed and xfind.attrnamed.