Fix media handling for html.applycss().
Add a method hasmedia() to html.style.Attrs.media and html.link.Attrs.media
(by putting it into their common baseclass), that checks whether a given
media type is in the attribute. Add checks for media type in the topmost
style loop, so that the media attributes on style and link elements are honored.