Add a class VPool, that can be used for "virtual class dispatching".
This will be the replacement for namespace inheritance, where
non-overwritten classes were automatically inherited and got a
new __ns__ attribute, which pointed back to the new namespace.

This new version doesn't create new element classes on the fly,
but and the reference back to the pool when the element classes
get instantiated.