Further work on the new parsers.
Rename low lower parser to LLParser (as we already have a class named Parser).

Add a method end() to the low level parser. This will be called after the
last call to feed() and can be used to clean up.

Remove the methods parsestring(), parseiter() and parsestream() from both
low level parsers (i.e. SGMLOPParser and ExpatParser) (they now live in
the Parser class).

Change the ExpatParser so that handler that could be directly called on
the application are wired to do so.

Add a transcode parameter to the ExpatParser constructor, if true the
input will be trancoded to utf-8 (This makes it possible to e.g.
parse shift-jis and other encodings with the ExpatParser).

The saxparser argument for the Parser constructor is now a parser
*instance* again. Parser objects will call begin() once, feed()
multiple times and end() once to do the parsing.

Added a global function parseiter() (More will follow).