Fix encoding handling in html.meta.convert()
When the meta content type was output with encoding==None,
html.meta.convert() couldn't handle it. Now it uses
publisher.getencoding() which returns the encoding in effect.

Publisher.getencoding() uses the encoding specified in the publisher.
If this is None it gets the encoding from the incremental encoder
(which sets its own encoding attribute once it has determined what
encoding to use). If the encoder hasn't determined the encoding
yet (e.g. because the meta element is the first one to be output,
use utf-8 (which will be what the encoder eventually will decide
to use too).