Fix parser bug. Fix .h file problem. Handle xmlns attribute in pool. Simplify URLAttr/PI interaction.
When parsing attributes that had an empty value, where drop. This is fixed
now.

distutils has dropped helpers.h from the source distributions. Fix this by
renaming the file to helpers_include.c.

Registerung module in a Pool object, now copies the xmlns attribute too (if it
exists and is a string).

"Fancy" URL attributes will now no longer be transformed in any way when
publishing.