Fix source attribute of sub templates.
The source must be the source of the top level template, so that offsets
in Location work. Add an endlocation attribute to sub templates, so that
sub templates can get at their source code by propert slicing.

Change Template.ul4load() and Template.ul4dump() so that attributes are
written in a reasonable order (which is different from the one inherited
from Block).