Introduce a rendering context for UL4 templates that contains the variables and will in the future contain a call stack to be able to reindent the output of a template when called from inside another template.
Rendering a template inside another template is now again done via the
<?render?> tag. The print and printx functions have been removed. The reason
for this is that what UL4 generator expressions output is ambiguous:
Iteratable items or the output of the template? With this change whether an
AST node does do output (and thus eval() is a generator) or not (and thus
eval() is a normal method) is specified with the class attribute output.
Output can only occur at the "top level" of the template, but not inside
expressions.