Rewrite many parts of pysql.
Calling eval() on the source of a pysql command no longer resolves var, loadstr and loadbytes objects. Instead those objects are now instances of subclasses of Command, which means executing a command must now recusrively execute nested subcommand.
Furthermore this means that include command now get executed by executing the include command instead of during loading.

The big advantage of this refactoring is: It is now possible to instantiate a command object directly. Printing command.source() will give the source that has to be put in a pysql file to recreate the pysql command. This should greatly simplify creating pysql command files.