Enhance source location handling for UL4 AST nodes.
Block nodes now have an additional attribute "stoppos" (and "pos" has been renamed to "startpos" or all nodes). "startpos" is the position of the start tag and "stoppos" is the position of the end tag.

Two attributes "startsource" and "stopsource" have been added. They return the source code of the start tag and the end tag.

The attributes "sourceprefix" and "sourcesuffix" have been renamed to
"startsourceprefix" and "startsourcesuffix".

(For non-block nodes "startsource" and "source" is the same.)

And a non-local template always has to start "startpos" for the start of the
source (i.e. slice(0, 0)) and a "stoppos" for the end of the source (i.e. slice(n, n) with n being the length of the source).