Fix the email logger.
Now both the parent and child process write the logs messages for the email into a file, and the parent process read both and writes the email.

Now we create a link "last_foo.siyphuslog" for each of the possible job results "foo".

Creating filename is no longer done with UL4 templates but with normal methods that return pathlib.Path objects.