Rewrite handling of delayed log files.
Delayed log messages were in the queue for both the parent and the child process. This is a problem if both log a non-delayed message. So we throw away queued messages in the parent process at the end of the run before cleanup up.

Also we don't recreate logs at the end, but only close specific log files (stdout and stderr are kept open).